refactor(assets): review folders and utils
