+++ apps/ai-dial-admin/src/components/PromptView/PromptProperties.tsx
