+++ apps/ai-dial-admin/src/utils/prompts/versions.ts
@@ -1,4 +1,3 @@
-import { DialPrompt } from '@/src/models/dial/prompt';
 import { PROMPT_NAME_VERSION_DELIMITER } from '@/src/constants/prompt';
 
 export const generateNameVersionForPrompt = (name: string, version: string) => {
@@ -61,38 +60,6 @@ export const checkNameVersionCombination = (
   version: string,
 ) => {
   return versionsPerName[name] && versionsPerName[name].includes(version);
-};
-
-export const getVersionsPerName = (data: DialPrompt[]) => {
-  const versionsPerName: Record<string, string[]> = {};
-
-  data.forEach((item) => {
-    const name = item.name as string;
-
-    if (!versionsPerName[name]) {
-      versionsPerName[name] = [];
-    }
-    versionsPerName[name].push(item.version);
-  });
-
-  Object.keys(versionsPerName).forEach((key) => {
-    versionsPerName[key] = versionsPerName[key].sort(compareVersions);
-  });
-
-  return versionsPerName;
-};
-
-export const filterLatestVersions = (data: DialPrompt[]) => {
-  const latestVersions: Record<string, DialPrompt> = {};
-
-  data?.forEach((item) => {
-    const name = item.name as string;
-    if (!latestVersions[name] || item.updateTime > latestVersions[name].updateTime) {
-      latestVersions[name] = item;
-    }
-  });
-
-  return Object.values(latestVersions);
 };
 
 const compareVersions = (version1: string, version2: string) => {
